+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Bug Fixes
 * Don't ignore fetch_size arguments to Statement constructors (github-151)
 * Allow disabling automatic paging on a per-statement basis when it's
   enabled by default for the session (PYTHON-93)
+* Raise ValueError when tuple query parameters for prepared statements
+  have extra items (PYTHON-98)
+* Correctly encode nested tuples and UDTs for non-prepared statements (PYTHON-100)
 
 Other
 -----